+++ src/main/java/com/team6/finalproject/comment/like/controller/CommentLikeController.java
@@ -1,14 +1,13 @@
-package com.team6.finalproject.comment.commentLike.controller;
+package com.team6.finalproject.comment.like.controller;
 
 import com.team6.finalproject.advice.custom.NotExistResourceException;
 import com.team6.finalproject.advice.custom.NotLikedYetException;
 import com.team6.finalproject.advice.custom.NotOwnedByUserException;
 import com.team6.finalproject.advice.custom.SelfLikeNotAllowedException;
-import com.team6.finalproject.comment.commentLike.service.CommentLikeService;
+import com.team6.finalproject.comment.like.service.CommentLikeService;
 import com.team6.finalproject.common.dto.ApiResponseDto;
 import com.team6.finalproject.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
