+++ src/main/java/com/team6/finalproject/comment/like/repository/CommentLikeRepository.java
@@ -1,6 +1,6 @@
-package com.team6.finalproject.comment.commentLike.repository;
+package com.team6.finalproject.comment.like.repository;
 
-import com.team6.finalproject.comment.commentLike.entity.CommentLike;
+import com.team6.finalproject.comment.like.entity.CommentLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
